refactor: moved some lsk logic to base class
